Fix duplicate code smells
Test file 'oidc-react.test.ts' had one duplicate test, which was removed.
Client, keycloak, and oidc-react had duplicate useClient and UseErrorDetection -hooks, which were moved to one place, client.ts. Duplicates in keycloak and oidc-react were removed and some shortcuts to client specific getters were added: 'useKeycloak' and 'useOidc'

getOrLoadUser were identical in both clients, but run in different context. Created a helper function that creates the function that is called  in the passed context.
