+++ src/clients/keycloak.ts
@@ -1,13 +1,10 @@
-import React, { useEffect, useState, useRef } from 'react';
 import Keycloak, { KeycloakTokenParsed } from 'keycloak-js';
 
 import {
   Client,
   ClientStatus,
-  ClientStatusId,
   User,
   ClientEvent,
-  ClientErrorObject,
   ClientError,
   createClient,
   ClientFactory,
@@ -16,7 +13,8 @@ import {
   getLocationBasedUri,
   ClientProps,
   getTokenUri,
-  FetchApiTokenOptions
+  FetchApiTokenOptions,
+  createClientGetOrLoadUserFunction
 } from './index';
 
 function getLocalStorageId(config: ClientProps, useType: string): string {
@@ -275,20 +273,11 @@ export function createKeycloakClient(): Client {
     return initPromise;
   };
 
-  const getOrLoadUser: Client['getOrLoadUser'] = () => {
-    const currentUser = getUser();
-    if (currentUser) {
-      return Promise.resolve(currentUser);
-    }
-    if (isInitialized()) {
-      return Promise.resolve(undefined);
-    }
-    return new Promise((resolve, reject) => {
-      init()
-        .then(() => resolve(getUser()))
-        .catch(e => reject(e));
-    });
-  };
+  const getOrLoadUser = createClientGetOrLoadUserFunction({
+    getUser,
+    isInitialized,
+    init
+  });
 
   const getApiAccessToken: Client['getApiAccessToken'] = async (
     options: FetchApiTokenOptions
@@ -335,75 +324,4 @@ export function getClient(): Client {
   }
   client = createKeycloakClient();
   return client;
-}
-
-export function useKeycloak(): Client {
-  const clientRef: React.Ref<Client> = useRef(getClient());
-  const clientFromRef: Client = clientRef.current as Client;
-  const [, setStatus] = useState<ClientStatusId>(clientFromRef.getStatus());
-  useEffect(() => {
-    const initClient = async (): Promise<void> => {
-      if (!clientFromRef.isInitialized()) {
-        await clientFromRef.getOrLoadUser().catch(e =>
-          clientFromRef.setError({
-            type: ClientError.INIT_ERROR,
-            message: e && e.toString()
-          })
-        );
-      }
-    };
-    const statusListenerDisposer = clientFromRef.addListener(
-      ClientEvent.STATUS_CHANGE,
-      status => {
-        setStatus(status as ClientStatusId);
-      }
-    );
-
-    initClient();
-    return (): void => {
-      statusListenerDisposer();
-    };
-  }, [clientFromRef]);
-  return clientFromRef;
-}
-
-export function useKeycloakErrorDetection(): ClientErrorObject {
-  const clientRef: React.Ref<Client> = useRef(getClient());
-  const clientFromRef: Client = clientRef.current as Client;
-  const [error, setError] = useState<ClientErrorObject>(undefined);
-  useEffect(() => {
-    let isAuthorized = false;
-    const statusListenerDisposer = clientFromRef.addListener(
-      ClientEvent.STATUS_CHANGE,
-      status => {
-        if (status === ClientStatus.AUTHORIZED) {
-          isAuthorized = true;
-        }
-        if (isAuthorized && status === ClientStatus.UNAUTHORIZED) {
-          setError({ type: ClientError.UNEXPECTED_AUTH_CHANGE, message: '' });
-          isAuthorized = false;
-        }
-      }
-    );
-
-    const errorListenerDisposer = clientFromRef.addListener(
-      ClientEvent.ERROR,
-      newError => {
-        setError(newError as ClientErrorObject);
-      }
-    );
-    const logoutListenerDisposer = clientFromRef.addListener(
-      ClientEvent.LOGGING_OUT,
-      (): void => {
-        isAuthorized = false;
-      }
-    );
-
-    return (): void => {
-      errorListenerDisposer();
-      statusListenerDisposer();
-      logoutListenerDisposer();
-    };
-  }, [clientFromRef]);
-  return error;
 }
