+++ src/clients/oidc-react.ts
@@ -12,14 +12,14 @@ import {
   ClientStatusId,
   User as ClientUser,
   ClientEvent,
-  ClientErrorObject,
   ClientError,
   createClient,
   ClientFactory,
   hasValidClientConfig,
   getClientConfig,
   getLocationBasedUri,
-  getTokenUri
+  getTokenUri,
+  createClientGetOrLoadUserFunction
 } from './index';
 
 let client: Client | null = null;
@@ -167,20 +167,11 @@ export function createOidcClient(): Client {
     return initPromise;
   };
 
-  const getOrLoadUser: Client['getOrLoadUser'] = () => {
-    const currentUser = getUser();
-    if (currentUser) {
-      return Promise.resolve(currentUser);
-    }
-    if (isInitialized()) {
-      return Promise.resolve(undefined);
-    }
-    return new Promise((resolve, reject) => {
-      init()
-        .then(() => resolve(getUser()))
-        .catch(e => reject(e));
-    });
-  };
+  const getOrLoadUser = createClientGetOrLoadUserFunction({
+    getUser,
+    isInitialized,
+    init
+  });
 
   const login: Client['login'] = () => {
     manager.signinRedirect();
@@ -294,70 +285,6 @@ export function getClient(): Client {
   }
   client = createOidcClient();
   return client;
-}
-
-export function useOidc(): Client {
-  const clientRef: React.Ref<Client> = useRef(getClient());
-  const clientFromRef: Client = clientRef.current as Client;
-  const [, setStatus] = useState<ClientStatusId>(clientFromRef.getStatus());
-  useEffect(() => {
-    const initClient = async (): Promise<void> => {
-      if (!clientFromRef.isInitialized()) {
-        await clientFromRef.getOrLoadUser().catch(e => {
-          clientFromRef.setError({
-            type: ClientError.INIT_ERROR,
-            message: e && e.toString()
-          });
-        });
-      }
-    };
-    const statusListenerDisposer = clientFromRef.addListener(
-      ClientEvent.STATUS_CHANGE,
-      status => {
-        setStatus(status as ClientStatusId);
-      }
-    );
-
-    initClient();
-    return (): void => {
-      statusListenerDisposer();
-    };
-  }, [clientFromRef]);
-  return clientFromRef;
-}
-
-export function useOidcErrorDetection(): ClientErrorObject {
-  const clientRef: React.Ref<Client> = useRef(getClient());
-  const clientFromRef: Client = clientRef.current as Client;
-  const [error, setError] = useState<ClientErrorObject>(undefined);
-  useEffect(() => {
-    let isAuthorized = false;
-    const statusListenerDisposer = clientFromRef.addListener(
-      ClientEvent.STATUS_CHANGE,
-      status => {
-        if (status === ClientStatus.AUTHORIZED) {
-          isAuthorized = true;
-        }
-        if (isAuthorized && status === ClientStatus.UNAUTHORIZED) {
-          setError({ type: ClientError.UNEXPECTED_AUTH_CHANGE, message: '' });
-          isAuthorized = false;
-        }
-      }
-    );
-
-    const errorListenerDisposer = clientFromRef.addListener(
-      ClientEvent.ERROR,
-      newError => {
-        setError(newError as ClientErrorObject);
-      }
-    );
-
-    return (): void => {
-      errorListenerDisposer();
-      statusListenerDisposer();
-    };
-  }, [clientFromRef]);
-  return error;
 }
 
 export const useOidcCallback = (): Client => {
